backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.35.0
celery[redis]==5.5.3
celery-sqlalchemy-scheduler==0.3.0
redis>=5,<7
psycopg[binary]==3.2.10
SQLAlchemy==2.0.43
alembic==1.16.5
pydantic==2.11.9
pydantic-settings==2.10.1
httpx==0.28.1
cryptography>=41.0.0
python-multipart>=0.0.6
jinja2>=3.1.2
prometheus-client==0.23.1
pyyaml==6.0.3
python-json-logger==3.3.0
python-dotenv==1.1.1
authlib==1.6.5
itsdangerous==2.2.0
attrs>=22.1.0
pytz
Mastodon.py>=1.8.0